Initial import of PostRR -- the PostgreSQL Round-Robin Extension.
For now, this is an early and highly experimental incomplete draft ;-)

The current version provides a new data type "RRTimeslice" implementing basic
round-robin features. It is defined by the length of the slice and the number
of slices before wrapping around, both specified using the type modifier. It
provides basic comparison operators and an operator class for the btree
indexing method.